[PP, PM-36] Moved method from rabbit to service
